+++ 07.bowling_object/frame.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+require './shot'
+
+class Frame
+  attr_reader :first_shot, :second_shot, :third_shot
+
+  def initialize(first_shot, second_shot, third_shot = nil)
+    @first_shot = Shot.new(first_shot)
+    @second_shot = Shot.new(second_shot)
+    @third_shot = Shot.new(third_shot)
+  end
+
+  def score
+    [first_shot.score, second_shot.score].sum
+    [first_shot.score, second_shot.score, third_shot.score].sum if third_shot
+  end
+end
